[Mb-Dynamic-Mapping] - Added Partitioner BasicBlock and Mapper IfmMapper; - Commit before changing the project to be able to profile;

+++ src/mb-dynamic-mapping/org/ancora/DmeFramework/Interfaces/MapperMonitor.java
@@ -24,7 +24,7 @@ package org.ancora.DmeFramework.Interfaces;
  */
 public interface MapperMonitor {
 
-   float getIlp();
+   double getIlp();
 
    int getLiveIn();
